Huge refactor, idk
