+++ src/common/list-layout/component.tsx
@@ -1,11 +1,10 @@
 import React, { ReactNode } from "react";
 import styled from "styled-components";
-import { STYLING_PROPS } from "./const";
 import {
   FLEXBOX_GAP_DATA,
   FLEXBOX_GAP_ENUM,
   FLEXBOX_GAP_TYPE,
-} from "../../theme/container-style";
+} from "../../theme/sizes";
 
 const Component: React.FC<{
   children?: ReactNode;
@@ -18,7 +17,7 @@ const Wrapper = styled.div`
   display: grid;
   max-width: 90%;
   margin: 12px auto;
-  gap: ${({ gap }: STYLING_PROPS) => FLEXBOX_GAP_DATA[gap]};
+  gap: ${(gap: FLEXBOX_GAP_TYPE) => FLEXBOX_GAP_DATA[gap]};
 `;
 
 export default Component;
