--- src/common/list-layout/const.tsx
@@ -1,5 +1,0 @@
-import { FLEXBOX_GAP_TYPE } from "../../theme/container-style";
-
-export interface STYLING_PROPS {
-  gap: FLEXBOX_GAP_TYPE;
-}
