+++ src/common/text-elem/component.tsx
@@ -1,7 +1,12 @@
 import React from "react";
 import { useTranslation } from "react-i18next";
 import styled from "styled-components";
-import { STYLING_PROPS } from "./const";
+import {
+  TEXT_COLOR_DATA,
+  TEXT_COLOR_ENUM,
+  TEXT_COLOR_TYPE,
+} from "../../theme/color";
+
 import {
   FONT_SIZE_DATA,
   FONT_SIZE_ENUM,
@@ -9,10 +14,7 @@ import {
   FONT_WEIGHT_DATA,
   FONT_WEIGHT_ENUM,
   FONT_WEIGHT_TYPE,
-  TEXT_COLOR_DATA,
-  TEXT_COLOR_ENUM,
-  TEXT_COLOR_TYPE,
-} from "../../theme/text-style";
+} from "../../theme/sizes";
 
 const Component: React.FC<{
   size?: FONT_SIZE_TYPE;
@@ -36,11 +38,11 @@ const Component: React.FC<{
 };
 
 const TextStyled = styled.p`
-  color: ${({ color }: STYLING_PROPS) => TEXT_COLOR_DATA[color]};
+  color: ${(color: TEXT_COLOR_TYPE) => TEXT_COLOR_DATA[color]};
   font-family: Manrope, sans-serif;
-  font-size: ${({ size }: STYLING_PROPS) => FONT_SIZE_DATA[size]};
+  font-size: ${(size: FONT_SIZE_TYPE) => FONT_SIZE_DATA[size]};
   margin: 0;
-  font-weight: ${({ weight }: STYLING_PROPS) => FONT_WEIGHT_DATA[weight]};
+  font-weight: ${(weight: FONT_WEIGHT_TYPE) => FONT_WEIGHT_DATA[weight]};
 `;
 
 export default Component;
