--- src/common/text-elem/const.tsx
@@ -1,12 +1,0 @@
-import React from "react";
-import {
-  FONT_SIZE_TYPE,
-  FONT_WEIGHT_TYPE,
-  TEXT_COLOR_TYPE,
-} from "../../theme/text-style";
-
-export interface STYLING_PROPS {
-  size: FONT_SIZE_TYPE;
-  color: TEXT_COLOR_TYPE;
-  weight: FONT_WEIGHT_TYPE;
-}
